Split out repeated test elements into a separate file
