+++ tests/babylonjs.test.mjs
@@ -1,4 +1,12 @@
 import { test, expect } from '@playwright/test';
+import {
+  injectIFrame,
+  EXAMPLE_URL,
+  EXAMPLE_IMAGE,
+  EXAMPLE_IMAGE2,
+  PREBID_LOAD_TEST_WAIT_INTERVAL,
+  PREBID_REFRESH_TEST_WAIT_INTERVAL
+} from './test-constants.mjs';
 
 test.beforeEach(async ({ page }) => {
   await page.goto('http://localhost:8080/tests/babylonjs/');
@@ -55,19 +63,9 @@ test.describe('Navigation', () => {
 });
 
 test.describe('Prebid', () => {
-  async function injectIFrame(page, url, image) {
-    await page.waitForFunction(() => document.querySelector('#zesty-div-medium-rectangle') != null);
-    await page.evaluate(([url, image]) => {
-      const iframe = document.createElement('iframe');
-      iframe.id = 'injected';
-      document.querySelector('#zesty-div-medium-rectangle').appendChild(iframe)
-      iframe.contentDocument.write(`<html><body><a href="${url}"><img src="${image}"></a></body></html>`);
-    }, [url, image]);
-  }
-
   test('Ad creative is loaded once bids is no longer null', async ({ page }) => {
-    await injectIFrame(page, 'https://www.example.com', 'https://picsum.photos/300/250');
-    await new Promise(res => setTimeout(res, 5000));
+    await injectIFrame(page, EXAMPLE_URL, EXAMPLE_IMAGE);
+    await new Promise(res => setTimeout(res, PREBID_LOAD_TEST_WAIT_INTERVAL));
     const img = await page.evaluate(
       () => window.scene.meshes[4].material.diffuseTexture.url
     );
@@ -75,11 +73,11 @@ test.describe('Prebid', () => {
   });
 
   test('A new ad creative is loaded after passing visibility check', async ({ page }) => {
-    await injectIFrame(page, 'https://www.example.com', 'https://picsum.photos/300/250');
-    await new Promise(res => setTimeout(res, 8000));
+    await injectIFrame(page, EXAMPLE_URL, EXAMPLE_IMAGE);
+    await new Promise(res => setTimeout(res, PREBID_REFRESH_TEST_WAIT_INTERVAL));
     await page.evaluate(() => document.querySelector('#injected').remove());
-    await injectIFrame(page, 'https://www.example.com', 'https://picsum.photos/300/300');
-    await new Promise(res => setTimeout(res, 8000));
+    await injectIFrame(page, EXAMPLE_URL, EXAMPLE_IMAGE2);
+    await new Promise(res => setTimeout(res, PREBID_REFRESH_TEST_WAIT_INTERVAL));
     const img = await page.evaluate(
       () => window.scene.meshes[4].material.diffuseTexture.url
     );
